Move the conversion from document to subject corpus entirely inside TFIDFBackend since nothing else uses it
