+++ tests/test_corpus.py
@@ -237,7 +237,7 @@ def test_docfile_is_empty(tmpdir):
     assert docs.is_empty()
 
 
-def test_combinedcorpus(tmpdir, subject_index):
+def test_combinedcorpus(tmpdir):
     docfile = tmpdir.join('documents.tsv')
     docfile.write("""Läntinen\t<http://www.yso.fi/onto/yso/p2557>
         Oulunlinnan\t<http://www.yso.fi/onto/yso/p7346>
@@ -247,7 +247,5 @@ def test_combinedcorpus(tmpdir, subject_index):
     corpus2 = annif.corpus.DocumentFile(str(docfile))
 
     combined = annif.corpus.CombinedCorpus([corpus1, corpus2])
-    combined.set_subject_index(subject_index)
 
     assert len(list(combined.documents)) == 6
-    assert len(list(combined.subjects)) == len(list(corpus1.subjects))
